Refactor Preferences component into modular structure
Extracted the state management logic to a custom hook (`usePreferences`) and
divided the presentation layer into smaller, domain-specific components:
- `PreferencesHeader`
- `PreferencesFooter`
- `TrainingPlanSection`
- `ModalitySections`
- `StyleSections`
- `PerformanceSections`

This significantly improves code readability and maintainability without altering functionality.
